build: Clarify pin mapping/don't override defines
Previously all pin mapping of ganged axes was an exercise left to the
user.

In ac96edd4 this behavior was changed to remove the burden from the user
through the use of sensible default behavior.

This extends that work by:

  - Clarifying the comments explaining the behavior (which were out of
    date in some cases)
  - Expands the explainations to avoid potential mis-configurations
  - Checking if the options if the following options were previously
    defined and only setting them in the event that they were not:
    - X_GANGED
    - X_AUTO_SQUARE
    - Y_GANGED
    - Y_AUTO_SQUARE
    - Z_GANGED
    - Z_AUTO_SQUARE

Users may now define *_GANGED and *_AUTO_SQUARE in `my_machine.h`
without raising compiler warnings or having the values re-defined in the
relevant map file.

All existing pin mappings were left as is to ensure there is no change
in behavior for existing users.